rename to inference
