+++ ai/inference/adapter.py
@@ -17,7 +17,7 @@ from typing import List, Optional
 
 from ai.proto.ai_model_pb2 import SingleModel
 
-from ai.runtime.types import ActionCommand, AdapterSpec, Observation
+from ai.inference.types import ActionCommand, AdapterSpec, Observation
 from config.proto import config_pb2
