+++ ros2/BUILD
@@ -56,12 +56,12 @@ ros2_py_binary(
         "//ai/models:sources",
         "//ai/models:manifests",
         "//ai/models/smolvla:requirements.lock",
-        "//ai/runtime:inference_launcher",
+        "//ai/inference:inference_launcher",
     ],
     main = "inference.py",
     set_up_ament = True,
     deps = [
-        "//ai/runtime:inference_launcher",
+        "//ai/inference:inference_launcher",
         "//config/proto:config_py_proto",
         "//ros2:node_runner_py",
         "//ros2/proto:trajectory_py_proto",
